create class CharClient, create class Program...
Program, the main program loop of the chat client system.
will later take arguments on startup if given to set the ip and port number to attempt connection to.
uses defaults if no args given.
now creates a client and attempts to connect to server.

ChatClient, handles the client, creates a Tcp client which attempts to connect to server at given ip and port.
Sets up socket stream reader and writer for communication.